requirements.txt
termcolor
pycryptodome
